backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pytest==8.0.2
httpx==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.3
pydantic==2.6.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
pdfplumber==0.11.0
PyMuPDF==1.26.7
pytesseract==0.3.13
pdf2image==1.17.0
Pillow>=10.0.0
python-multipart==0.0.9
